Changing two tests and changing one test to another class
