+++ app/src/androidTest/java/ch/epfl/sweng/runpharaa/TrackPropertiesActivityTest.java
@@ -89,6 +89,7 @@ public class TrackPropertiesActivityTest extends TestInitLocation {
         launchWithExtras(t1);
         onView(withId(R.id.buttonLikeID)).perform(click());
         withId(R.id.trackLikesID).matches(withText("1"));
+        sleep(3000);
 
     }
 
@@ -98,6 +99,7 @@ public class TrackPropertiesActivityTest extends TestInitLocation {
         launchWithExtras(t1);
         onView(withId(R.id.buttonFavoriteID)).perform(click());
         withId(R.id.trackFavouritesID).matches(withText("1"));
+        sleep(4000);
     }
 
     @Test
